adapting edit mode on profile section rather than page rendering
